move fn to http-api and add jsdoc
